Make price a Double

+++ Components/ViewComponents/Price/PriceComponent.swift
@@ -3,12 +3,12 @@ import Foundation
 struct PriceComponent: Component {
 
     let id: String
-    let price: Int
+    let price: Double
     let locale: Locale
     let accessibilityPrefix: String
     let status: String?
 
-    init(id: String = UUID().uuidString, price: Int, locale: Locale, accessibilityPrefix: String, status: String? = nil) {
+    init(id: String = UUID().uuidString, price: Double, locale: Locale, accessibilityPrefix: String, status: String? = nil) {
         self.id = id
         self.price = price
         self.locale = locale
